Move watchlist api tests to seperate module
