+++ piker/watchlists.py
@@ -6,11 +6,10 @@ from .log import get_logger
 
 log = get_logger(__name__)
 
+
 def write_sorted_json(watchlist, path):
     for key in watchlist:
-        s = set(watchlist[key])
-        watchlist[key] = list(s)
-        watchlist[key].sort()
+        watchlist[key] = sorted(list(set(watchlist[key])))
     with open(path, 'w') as f:
         json.dump(watchlist, f, sort_keys=True)
 
@@ -24,29 +23,23 @@ def make_config_dir(dir_path):
 def ensure_watchlists(file_path):
     mode = 'r' if os.path.isfile(file_path) else 'w'
     with open(file_path, mode) as f:
-        data = json.load(f) if not os.stat(file_path).st_size == 0 else {}
-    return data
+        return json.load(f) if not os.stat(file_path).st_size == 0 else {}
 
 
-def load_watchlists(watchlist, path):
-    watchlist = json.loads(watchlist)
-    write_sorted_json(watchlist, path)
-
-
-def new_group(name, watchlist, path):
-    watchlist.setdefault(name, [])
-    write_sorted_json(watchlist, path)
+def write_watchlists(watchlist, path):
+    write_sorted_json(json.loads(watchlist), path)
 
 
 def add_ticker(name, ticker_name, watchlist, path):
-    if name in watchlist:
-        watchlist[name].append(str(ticker_name).upper())
+    watchlist.setdefault(name, []).append(str(ticker_name).upper())
     write_sorted_json(watchlist, path)
 
 
 def remove_ticker(name, ticker_name, watchlist, path):
     if name in watchlist:
         watchlist[name].remove(str(ticker_name).upper())
+        if watchlist[name] == []:
+            del watchlist[name]
     write_sorted_json(watchlist, path)
